% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgtap_functions.r
\name{randomstring}
\alias{randomstring}
\title{Generates a random string of characters.}
\usage{
randomstring(n = 1, lenght = 12)
}
\arguments{
\item{solution.out}{It's the name (or directory?) where you put the solution}

\item{csv.out}{name of the csv file you want to read from.}

\item{header}{header which is read.}
}
\description{
Generates a random string of characters.
}
\examples{
Examples
randomstring()
}

