% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgtap_functions.r
\name{extractvar}
\alias{extractvar}
\title{Extract variables for GEMPACK solution files.}
\usage{
extractvar(solution.dir, solution.name, var.map, solution.out)
}
\arguments{
\item{solution.dir}{Directory where the solution file is stored.}

\item{solution.name}{Name of the solution, including suffix}

\item{var.map}{A GEMPACK variable map to be used with "sltoht". Need to add a reference to the GTAP manual.}

\item{solution.out}{It's the name (or directory?) where you put the solution}
}
\description{
Extract variables for GEMPACK solution files.
}
\examples{
Examples
extractvar()
}

