% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgtap_functions.r
\name{readsol}
\alias{readsol}
\title{Read specific header of solution files, conver them into CSV files,
and read them in R.}
\usage{
readsol(solution.out, csv.out, header)
}
\arguments{
\item{solution.out}{It's the name (or directory?) where you put the solution}

\item{csv.out}{name of the csv file you want to read from.}

\item{header}{header which is read.}
}
\description{
Read specific header of solution files, conver them into CSV files,
and read them in R.
}
\examples{
Examples
extractvar_function()
}

